refactor: move io functions to dedicated modules
