+++ src/shacl_api/server.py
@@ -1,16 +1,15 @@
 from enum import Enum
 import os
-import shutil
-import subprocess
 import tempfile
-from typing import Annotated, IO, Optional, Union
+from typing import Annotated, Union
 
 from fastapi import FastAPI, File, Header, UploadFile
 from fastapi.responses import FileResponse
 from pydantic import BaseModel
-from rdflib import Graph
 
 from . import __version__
+from .io import convert_rdf_file
+from .run import ShaclCommand, run_shacl
 
 description = """
 
@@ -77,58 +76,11 @@ class FormatHeaders(BaseModel):
     """Common headers for RDF formats."""
     accept: RdfMimeType = RdfMimeType.jsonld
 
-def convert_rdf_file(
-    in_file: IO[bytes],
-    to_file: IO[bytes],
-    in_format: str="turtle",
-    to_format: str="turtle",
-):
-    """Convert RDF data to Turtle format if not already in Turtle."""
-    if in_format == to_format:
-        # NOTE: may want to symlink instead to remove overhead
-        shutil.copyfileobj(in_file, to_file)
-    else:
-        # WARN: This loads the file in memory
-        g = Graph()
-        g.parse(in_file, format=in_format)
-        to_file.write(g.serialize(format=to_format).encode())
-    to_file.seek(0)
 
-class ShaclCommand(str, Enum):
-    """Supported SHACL commands."""
-    validate = "shaclvalidate.sh"
-    infer = "shaclinfer.sh"
+def _wrap_shacl(mode: ShaclCommand, data: UploadFile, shapes: Union[UploadFile, str, None], headers: FormatHeaders):
+    """Wrap SHACL command execution and RDF format conversion for use in endpoints."""
 
-def run_shacl(mode: ShaclCommand, data_path: str, shapes_path: str) -> IO[bytes]:
-    report_file = tempfile.NamedTemporaryFile(delete=False, delete_on_close=False)
-    _ = subprocess.run(
-        [mode.value, "-datafile", data_path, "-shapesfile", shapes_path],
-        stdout=report_file,
-    )
-    report_file.seek(0)
-    
-    return report_file
-
-
-@app.post("/validate", )
-def validate(
-    data: Annotated[
-        UploadFile,
-        File(description="RDF file with instance data to validate.")
-    ],
-    shapes: Annotated[
-        Union[UploadFile, str, None],
-        File(description="SHACL shapes file. Default: None.")
-    ] = None,
-    headers: Annotated[
-        FormatHeaders,
-        Header(description="Request headers for RDF format")] = FormatHeaders(),
-):
-    """Validate RDF instance data against SHACL shapes.
-    If shapes are omitted, the default shapes file from the server is used.
-    Supported file formats are rdf-xml, turtle, json-ld and ntriples.
-    """
-
+    # Fall back on server shapes if none are provided
     if shapes is None or isinstance(shapes, str):
         shapes_source = open(SHAPES_PATH, 'rb')
         shapes_format = "turtle"
@@ -153,7 +105,7 @@ def validate(
         )
 
         # Run validation
-        report_ttl = run_shacl(ShaclCommand.validate, data_ttl.name, shapes_ttl.name)
+        report_ttl = run_shacl(mode, data_ttl.name, shapes_ttl.name)
 
     # Convert report to requested format
     report_file = tempfile.NamedTemporaryFile(delete=False, delete_on_close=False)
@@ -167,8 +119,8 @@ def validate(
     # NOTE: FastAPI automatically deletes the response file
     return FileResponse(report_file.name, media_type=headers.accept)
 
-@app.post("/infer", )
-def infer(
+@app.post("/validate", )
+def validate(
     data: Annotated[
         UploadFile,
         File(description="RDF file with instance data to validate.")
@@ -181,16 +133,27 @@ def infer(
         FormatHeaders,
         Header(description="Request headers for RDF format")] = FormatHeaders(),
 ):
+    """Validate RDF instance data against SHACL shapes.
+    If shapes are omitted, the default shapes file from the server is used.
+    Supported file formats are rdf-xml, turtle, json-ld and ntriples.
+    """
+    return _wrap_shacl(ShaclCommand.validate, data, shapes, headers)
+
+
+@app.post("/infer", )
+def infer(
+    data: Annotated[
+        UploadFile,
+        File(description="RDF file with instance data on which to run inference.")
+    ],
+    shapes: Annotated[
+        Union[UploadFile, str, None],
+        File(description="SHACL shapes file. Default: None.")
+    ] = None,
+    headers: Annotated[
+        FormatHeaders,
+        Header(description="Request headers for RDF format")] = FormatHeaders(),
+):
     """Runs SHACL inference on RDF instance data using provided shapes. If shapes are omitted, the default shapes file from the server is used. Supported file formats are rdf-xml, turtle, json-ld and ntriples."""
 
-    output = subprocess.run(
-        ["shaclinfer.sh", "-datafile", "datafile.ttl", "-shapesfile", "shapesfile.ttl"],
-        stdout=subprocess.PIPE,
-    )
-
-    os.remove("datafile.ttl")
-    os.remove("shapesfile.ttl")
-    with open("validationTest.ttl", "wb") as f:
-        f.write(output.stdout)
-
-    return {"output": output.stdout}
+    return _wrap_shacl(ShaclCommand.infer, data, shapes, headers)
